Ajout des conditions lorsque l'objet sélectionné est null (clic droit qui ne sélectionne pas un objet)

+++ src/main/java/fr/kyo/crkf/controller/GestionClassificationController.java
@@ -85,25 +85,29 @@ public class GestionClassificationController {
 
     @FXML
     private void remove(){
-        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
-        alert.setTitle("Supprimer");
-        alert.setHeaderText("Voulez vous vraiment supprimer cet element?");
-        Optional<ButtonType> result = alert.showAndWait();
-        if(result.isPresent() && result.get() == ButtonType.OK){
-            if(classificationTable.getSelectionModel().getSelectedItem().getNumberFamilles().getValue() > 0){
-                alert.close();
-                Alert alert1 = new Alert(Alert.AlertType.ERROR);
-                alert1.setTitle("Erreur");
-                alert1.setHeaderText("Il y a eu une erreur lors de la suppresion de la classification.\nIl est impossible de supprimer une classification qui a des familles associées");
-                alert1.showAndWait();
-            }else{
-                DAOFactory.getClassificationDAO().delete(classificationTable.getSelectionModel().getSelectedItem());
-                filter();
+
+        if (classificationTable.getSelectionModel().getSelectedItem() != null){
+            Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+            alert.setTitle("Supprimer");
+            alert.setHeaderText("Voulez vous vraiment supprimer cet element?");
+            Optional<ButtonType> result = alert.showAndWait();
+            if(result.isPresent() && result.get() == ButtonType.OK){
+                if(classificationTable.getSelectionModel().getSelectedItem().getNumberFamilles().getValue() > 0){
+                    alert.close();
+                    Alert alert1 = new Alert(Alert.AlertType.ERROR);
+                    alert1.setTitle("Erreur");
+                    alert1.setHeaderText("Il y a eu une erreur lors de la suppresion de la classification.\nIl est impossible de supprimer une classification qui a des familles associées");
+                    alert1.showAndWait();
+                }else{
+                    DAOFactory.getClassificationDAO().delete(classificationTable.getSelectionModel().getSelectedItem());
+                    filter();
+                }
             }
         }
     }
     @FXML
     private void update(){
+        if (classificationTable.getSelectionModel().getSelectedItem() != null)
         applicationCRKF.openUpdateClassificationModal(this, classificationTable.getSelectionModel().getSelectedItem());
     }
 }
