+++ src/main/java/fr/kyo/crkf/controller/ville/GestionVilleController.java
@@ -46,6 +46,7 @@ public class GestionVilleController {
     }
     @FXML
     private void openUpdateModal(){
+        if (villeTable.getSelectionModel().getSelectedItem() != null)
         applicationCRKF.openModalUpdateVille(this, villeTable.getSelectionModel().getSelectedItem());
     }
 
@@ -82,13 +83,15 @@ public class GestionVilleController {
     }
     @FXML
     private void remove(){
-        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
-        alert.setTitle("Supprimer");
-        alert.setHeaderText("Voulez-vous vraiment supprimer cet element?");
-        Optional<ButtonType> result = alert.showAndWait();
-        if(result.isPresent() && result.get() == ButtonType.OK)
-            DAOFactory.getVilleDAO().delete(villeTable.getSelectionModel().getSelectedItem());
-        filter();
+        if (villeTable.getSelectionModel().getSelectedItem() != null){
+            Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+            alert.setTitle("Supprimer");
+            alert.setHeaderText("Voulez-vous vraiment supprimer cet element?");
+            Optional<ButtonType> result = alert.showAndWait();
+            if(result.isPresent() && result.get() == ButtonType.OK)
+                DAOFactory.getVilleDAO().delete(villeTable.getSelectionModel().getSelectedItem());
+            filter();
+        }
     }
 
     @FXML
